adults-bot copy/requirements.txt
aiogram
firebase_admin
google_api_python_client
protobuf
